GitHub
I have developed a Deep Learning project called AI-Powered Resume Screening System.  

The main benefits of this system are that when a company posts job vacancies, a large number of candidates upload their CVs or resumes. Manually matching each CV or resume with the company’s job criteria is extremely difficult and time-consuming.  

My software Aanalyzes all the CVs and Resumes*and Automatically Saves the ones that match the company’s eligibility criteria into a specific folder.  
Meanwhile, the CVs or resumes that Do Not Match the criteria are Automatically Deleted.  

This not only Saves a lot of time but also Benefits the Company by streamlining the recruitment process.